requirements.txt
dotenv
tqdm
requests
pathlib
pyyaml
jinja2
psycopg2
pillow